Generate contract from module
