+++ apps/.gitignore
@@ -8,6 +8,3 @@
 !.yarn/sdks
 !.yarn/versions
 node_modules
-
-# apps
-dist
