+++ apps/voting/.gitignore
@@ -1,0 +1,2 @@
+dist
+src/__generated__
