+++ apps/voting/src/pages/Home/Home.tsx
@@ -1,0 +1,3 @@
+export default function Home() {
+    return <>Home</>;
+}
