+++ apps/voting/src/shared/election-contract.ts
@@ -1,0 +1,1 @@
+export * as ElectionContract from '../__generated__/election-contract/module_concordium_governance_committee_election';
